requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
easyocr==1.7.2
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
fasttext==0.9.3
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
pip-autoremove==0.10.0
protobuf==4.25.5
pybind11==2.13.6
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-bidi==0.6.3
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
typer==0.13.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.0
Werkzeug==3.1.3
wrapt==1.17.0
